:recycle: Code refactor and submodule renaming
